mise en place de l architecture  des page du site avec le portfolio en page de garde
