+++ front/src/stores/router.js
@@ -1,7 +1,7 @@
 import { writable } from 'svelte/store';
 
 // On lit la valeur du localStorage au démarrage
-const initialPage = localStorage.getItem('currentPage') || 'acceuil';
+const initialPage = localStorage.getItem('currentPage') || 'acceuilPortfolioBis';
 
 // On crée le store avec la valeur initiale
 export const currentPage = writable(initialPage);
